pyproject.toml
[tool.pdm]
[tool.pdm.dev-dependencies]
lint = [
    "flake8",
    "black"
]
test = ["pytest", "pytest-cov"]
dev = [
    "-e file:///${PROJECT_ROOT}/.",
]

[project]
name = "dolly"
version = "0.0.1"
description = "SURF port of dolly"
authors = [
    {name = "robogast", email = "robertjan.schlimbach@gmail.com"},
]
dependencies = [
    "click==8.0.3",
    "watchdog>=2.1.9",
    "accelerate>=0.18.0",
    "datasets==2.8.0",
    "transformers[torch]>=4.25.1",
    "deepspeed>=0.8.3",
    "pytorch-lightning>=2.0.1",
    "torch==2.1.0.dev20230403+cu118",
    "torchvision==0.16.0.dev20230403+cu118",
    "hydra-core==1.3.2",
    "tensorboardX>=2.6",
    "setuptools>=67.6.1", # Otherwise I get a "ModuleNotFoundError: No module named 'pkg_resources'" error from ...
]

requires-python = ">=3.10"
license = {text = "Apache License 2.0"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[[tool.pdm.source]]
type = "find_links"
url = "https://download.pytorch.org/whl/nightly/torch/"
name = "torch-src"

[[tool.pdm.source]]
type = "find_links"
url = "https://download.pytorch.org/whl/nightly/torchvision/"
name = "torchvision-src"

[[tool.pdm.source]]
type = "find_links"
url = "https://download.pytorch.org/whl/nightly/pytorch-triton"
name = "torch-triton-src"
